requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-hashers-passlib==0.4
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
passlib==1.7.2
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
RestAuth==0.6.3
RestAuthCommon==0.7.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
South==1.0.2
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
